+++ contracts/ProtocolControl.sol
@@ -8,7 +8,6 @@ import "@openzeppelin/contracts/access/AccessControl.sol";
 import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 
 contract ProtocolControl is AccessControl {
-    
     /// @dev Admin role for protocol.
     bytes32 public constant PROTOCOL_ADMIN = keccak256("PROTOCOL_ADMIN");
     /// @dev Admin role for NFTLabs.
@@ -21,26 +20,41 @@ contract ProtocolControl is AccessControl {
     address public nftlabsTreasury;
 
     /// @dev Pack protocol module names.
-    enum ModuleType { Coin, NFT, Pack, Market, Other }
+    enum ModuleType {
+        Coin,
+        NFT,
+        Pack,
+        Market,
+        Other
+    }
 
     /// @dev Module ID => Module address.
     mapping(bytes32 => address) public modules;
     /// @dev Module ID => Module type.
     mapping(bytes32 => ModuleType) public moduleType;
     /// @dev Module type => Num of modules of that type.
-    mapping(uint => uint) public numOfModuleType;
+    mapping(uint256 => uint256) public numOfModuleType;
 
     /// @dev Market fees
     uint256 public constant MAX_BPS = 10000; // 100%
-    uint public marketFeeBps;
+    uint256 public marketFeeBps;
 
     /// @dev Events.
-    event ProtocolInitialized(address pack, bytes32 packModuleId, address market, bytes32 marketModuleId, address coin, bytes32 coinModuleId, address nft, bytes32 nftModuleId);
+    event ProtocolInitialized(
+        address pack,
+        bytes32 packModuleId,
+        address market,
+        bytes32 marketModuleId,
+        address coin,
+        bytes32 coinModuleId,
+        address nft,
+        bytes32 nftModuleId
+    );
     event ModuleInitialized(bytes32 moduleId, address module);
-    event ModuleUpdated(bytes32 indexed moduleId, address indexed module, uint indexed moduleType);
+    event ModuleUpdated(bytes32 indexed moduleId, address indexed module, uint256 indexed moduleType);
     event FundsTransferred(address asset, address to, uint256 amount);
     event SystemPaused(bool isPaused);
-    event MarketFeeBps(uint marketFeeBps);
+    event MarketFeeBps(uint256 marketFeeBps);
 
     /// @dev Check whether the caller is a protocol admin
     modifier onlyProtocolAdmin() {
@@ -54,11 +68,7 @@ contract ProtocolControl is AccessControl {
         _;
     }
 
-    constructor(
-        address _admin,
-        address _nftlabs
-    ) {
-
+    constructor(address _admin, address _nftlabs) {
         nftlabsTreasury = _nftlabs;
 
         _setupRole(NFTLABS, _nftlabs);
@@ -69,30 +79,37 @@ contract ProtocolControl is AccessControl {
     }
 
     /// @dev Iniializes all components of the protocol.
-    function initializeProtocol(address _coin, address _nft, address _pack, address _market) external onlyProtocolAdmin {
+    function initializeProtocol(
+        address _coin,
+        address _nft,
+        address _pack,
+        address _market
+    ) external onlyProtocolAdmin {
         require(
-            numOfModuleType[uint(ModuleType.Pack)] == 0 
-            && numOfModuleType[uint(ModuleType.Market)] == 0
-            && numOfModuleType[uint(ModuleType.NFT)] == 0
-            && numOfModuleType[uint(ModuleType.Coin)] == 0,
+            numOfModuleType[uint256(ModuleType.Pack)] == 0 &&
+                numOfModuleType[uint256(ModuleType.Market)] == 0 &&
+                numOfModuleType[uint256(ModuleType.NFT)] == 0 &&
+                numOfModuleType[uint256(ModuleType.Coin)] == 0,
             "Protocol Control: already initialized."
         );
 
-
-        bytes32 packModuleId = addModule(_pack, uint(ModuleType.Pack));
-        bytes32 marketModuleId = addModule(_market, uint(ModuleType.Market));
-        bytes32 coinModuleId = addModule(_nft, uint(ModuleType.NFT));
-        bytes32 nftModuleId = addModule(_coin, uint(ModuleType.Coin));
+        bytes32 packModuleId = addModule(_pack, uint256(ModuleType.Pack));
+        bytes32 marketModuleId = addModule(_market, uint256(ModuleType.Market));
+        bytes32 coinModuleId = addModule(_nft, uint256(ModuleType.NFT));
+        bytes32 nftModuleId = addModule(_coin, uint256(ModuleType.Coin));
 
         emit ProtocolInitialized(_pack, packModuleId, _market, marketModuleId, _coin, coinModuleId, _nft, nftModuleId);
     }
 
     /// @dev Lets a protocol admin change the address of a module of the protocol.
-    function addModule(address _newModuleAddress, uint _moduleType) public onlyProtocolAdmin returns (bytes32 moduleId)  {
-        
+    function addModule(address _newModuleAddress, uint256 _moduleType)
+        public
+        onlyProtocolAdmin
+        returns (bytes32 moduleId)
+    {
         moduleId = keccak256(abi.encodePacked(numOfModuleType[_moduleType]));
         numOfModuleType[_moduleType] += 1;
-        
+
         modules[moduleId] = _newModuleAddress;
 
         emit ModuleUpdated(moduleId, _newModuleAddress, _moduleType);
@@ -102,11 +119,11 @@ contract ProtocolControl is AccessControl {
     function updateModule(bytes32 _moduleId, address _newModuleAddress) external onlyProtocolAdmin {
         modules[_moduleId] = _newModuleAddress;
 
-        emit ModuleUpdated(_moduleId, _newModuleAddress, uint(moduleType[_moduleId]));
+        emit ModuleUpdated(_moduleId, _newModuleAddress, uint256(moduleType[_moduleId]));
     }
 
     /// @dev Lets a nftlabs admin change the market fee basis points.
-    function updateMarketFeeBps(uint _newFeeBps) external onlyNftlabsAdmin {
+    function updateMarketFeeBps(uint256 _newFeeBps) external onlyNftlabsAdmin {
         marketFeeBps = _newFeeBps;
 
         emit MarketFeeBps(_newFeeBps);
